Adds Mute follow-up links and two more settings
1. Mute Link:

As of this release there's a link added in each
follow-up message that allow the user the capability
to stop following the comment thread.
There's a new signal 'signals.comment_thread_muted'
send every time a user clicks on the mute link. The signal
won't disable the user request but may be useful to track
how users disable follow-up messages on different circumstances.

2. setting COMMENTS_XTD_SEND_HTML_EMAIL:

Boolean setting set to True by default. When set to False
emails will be sent only in text mode.

3. setting COMMENTS_XTD_THREADED_EMAILS:

Boolean setting set to True by default. When set to False
emails will be sent using the main thread under which the
current request is running. That leaves the possibility to
use an in-house Celery application to handle high volume
traffic of emails. Or a 3rd party app like django-celery-email.

Fixes #10, fixes #13.