Fixes issue #229 about REST API post comment failing on honeypot field.
From version 2.7.0, posting a comment using the REST API failed to
process the honeypot field. This commit resolves the issue. There are
two ways to post comments using the REST API, one consist of providing
the honeypot, timestamp and security_hash fields, and the other consist
of authorizing the request (See the Web API in the docs).