New comment-reaction controller to replace all comment flags.
This commit removes the like/dislike URLs and controllers and adds an
API URL to handle customizable comment reactions. The CommentReaction
model comes to replace the usage of the CommentFlag model.
CommentReaction instances keep counters for each comment/reaction pair,
so it is easier to find out what comments receive more attention.